backend/requirements.txt
asgiref==3.8.1
click==8.2.1
colorama==0.4.6
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==2.3.2
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.5.0
tzdata==2025.2
